# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Manager API",
        "version": "1.0.0",
        "description": "A simple book management API"
    },
    "components": {
        "schemas": {
            "Book": {
                "type": "object",
                "properties": {
                    "isbn": {
                        "type": "string",
                        "example": "B001"
                    },
                    "title": {
                        "type": "string",
                        "example": "Clean Code"
                    },
                    "author": {
                        "type": "string",
                        "example": "Robert C. Martin"
                    },
                    "year": {
                        "type": "integer",
                        "example": 2008
                    },
                    "category": {
                        "type": "string",
                        "example": "Programming"
                    }
                },
                "required": ["isbn", "title", "author", "year", "category"]
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "paths": {
        "/books": {
            "get": {
                "summary": "Get all books",
                "responses": {
                    "200": {
                        "description": "List of all books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new book",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{isbn}": {
            "get": {
                "summary": "Get a book by ISBN",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "B001"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a book",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a book",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted successfully"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
        "/books/search": {
            "get": {
                "summary": "Search books by category",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Programming"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Category is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/sort": {
            "get": {
                "summary": "Sort books by title or year",
                "parameters": [
                    {
                        "name": "by",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": ["title", "year"]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": ["asc", "desc"]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sorted results"
                    }
                }
            }
        }
    }
}